# Optimize hasDoubleWildcard